Extract Table viz definition into a typed definition file
